refactored structure
