+++ javascript/index.js
@@ -1,5 +1,10 @@
+/**
+ * Initializes the Snake Game application.
+ * Dynamically imports the game module and starts the game on the canvas element.
+ * @returns {void}
+ */
 function initApp() {
-    import("./game/game.js").then(({ createGame }) => {
+    import("./core/game.js").then(({ createGame }) => {
         const canvas = document.getElementById("canvas");
         createGame(canvas).initialize();
     });
